+++ quint/src/compileToTlaplus.ts
@@ -21,12 +21,16 @@ import { ApalacheResult, ServerEndpoint, connect } from './apalache'
  * @param serverEndpoint
  *   a server endpoint
  *
+ * @param apalacheVersion
+ *  the version of Apalache to use if there is no active server connection
+ *
  * @param parseDataJson the flattened, analyzed, parse data, in as a json string
  *
  * @returns right(tlaString) if parsing succeeds, or left(err) explaining the failure
  */
 export async function compileToTlaplus(
   serverEndpoint: ServerEndpoint,
+  apalacheVersion: string,
   parseDataJson: string,
   verbosityLevel: number
 ): Promise<ApalacheResult<string>> {
@@ -39,6 +43,6 @@ export async function compileToTlaplus(
       },
     },
   }
-  const connectionResult = await connect(serverEndpoint, verbosityLevel)
+  const connectionResult = await connect(serverEndpoint, apalacheVersion, verbosityLevel)
   return connectionResult.asyncChain(conn => conn.tla(config))
 }